Restructure Tlaib Income Security Package to Economic Dignity for All Agenda
- Rename income_security_package to economic_dignity_for_all_agenda
- Remove Baby Bonus Act and BOOST Act components (will be modeled in app using existing parameters)
- Isolate End Child Poverty Act as sole structural reform component
- Rename implementation to edaa_end_child_poverty_act.py to differentiate from original version
- Refactor tests to focus on ECPA functionality
- Update imports in reforms.py and tlaib/__init__.py
- Net result: -1,367 lines (cleaner, focused implementation)

This change eliminates household benefit confusion and cross-contamination between reforms.
