--- policyengine_us/parameters/gov/contrib/congress/tlaib/income_security_package/baby_bonus_act/age_limit.yaml
@@ -1,8 +1,0 @@
-description: Representative Rashida Tlaib's Baby Bonus Act provides a credit to filers with children above this age limit.
-metadata:
-  unit: year
-  period: year
-  label: Baby bonus age limit
-  reference: placeholder - bill not yet introduced
-values:
-  2026-01-01: 1
