--- policyengine_us/parameters/gov/contrib/congress/tlaib/income_security_package/baby_bonus_act/in_effect.yaml
@@ -1,8 +1,0 @@
-description: Representative Rashida Tlaib's Baby Bonus Act is in effect when this is true.
-metadata:
-  unit: bool
-  period: year
-  # reference: placeholder - bill not yet introduced
-  label: Baby Bonus Act in effect
-values:
-  0000-01-01: false
