--- policyengine_us/parameters/gov/contrib/congress/tlaib/income_security_package/baby_bonus_act/min_birth_year.yaml
@@ -1,8 +1,0 @@
-description: Representative Rashida Tlaib's Baby Bonus Act sets the minimum birth year for eligibility.
-metadata:
-  unit: year
-  period: year
-  label: Baby bonus minimum birth year
-  reference: placeholder - bill not yet introduced
-values:
-  2026-01-01: 2026
