--- policyengine_us/parameters/gov/contrib/congress/tlaib/income_security_package/boost_act/in_effect.yaml
@@ -1,8 +1,0 @@
-description: Representative Rashida Tlaib's BOOST Act is in effect when this is true.
-metadata:
-  unit: bool
-  period: year
-  # reference: placeholder - bill not yet introduced
-  label: BOOST Act in effect
-values:
-  0000-01-01: false
