--- policyengine_us/parameters/gov/contrib/congress/tlaib/income_security_package/boost_act/max_age.yaml
@@ -1,8 +1,0 @@
-description: Representative Rashida Tlaib's BOOST Act sets this maximum age for eligibility.
-metadata:
-  unit: year
-  period: year
-  # reference: placeholder - bill not yet introduced
-  label: BOOST Act maximum age
-values:
-  2026-01-01: 67
