--- policyengine_us/parameters/gov/contrib/congress/tlaib/income_security_package/boost_act/min_age.yaml
@@ -1,8 +1,0 @@
-description: Representative Rashida Tlaib's BOOST Act sets this minimum age for eligibility.
-metadata:
-  unit: year
-  period: year
-  # reference: placeholder - bill not yet introduced
-  label: BOOST Act minimum age
-values:
-  2026-01-01: 19
