--- policyengine_us/parameters/gov/contrib/congress/tlaib/income_security_package/boost_act/tax/rate.yaml
@@ -1,8 +1,0 @@
-description: Representative Rashida Tlaib's BOOST Act applies this supplemental tax rate on income above certain thresholds.
-metadata:
-  unit: /1
-  period: year
-  # reference: placeholder - bill not yet introduced
-  label: BOOST Act tax rate
-values:
-  2026-01-01: 0.025
