+++ policyengine_us/reforms/reforms.py
@@ -28,8 +28,9 @@ from .harris.rent_relief_act.rent_relief_tax_credit import (
 )
 from .congress.tlaib import (
     create_end_child_poverty_act_reform,
-    income_security_package,
-    create_income_security_package_reform,
+)
+from .congress.tlaib.economic_dignity_for_all_agenda import (
+    create_end_child_poverty_act_reform as create_edaa_end_child_poverty_act_reform,
 )
 from .congress.tlaib.boost import (
     create_boost_middle_class_tax_credit_reform,
@@ -170,7 +171,7 @@ def create_structural_reforms_from_parameters(parameters, period):
     end_child_poverty_act = create_end_child_poverty_act_reform(
         parameters, period
     )
-    income_security_package_reform = create_income_security_package_reform(
+    edaa_end_child_poverty_act = create_edaa_end_child_poverty_act_reform(
         parameters, period
     )
     boost_middle_class_tax_credit = (
@@ -288,7 +289,7 @@ def create_structural_reforms_from_parameters(parameters, period):
         middle_class_tax_credit,
         rent_relief_tax_credit,
         end_child_poverty_act,
-        income_security_package_reform,
+        edaa_end_child_poverty_act,
         boost_middle_class_tax_credit,
         mn_walz_hf1938,
         or_rebate_state_tax_exempt,
